Compile parsers separately
