+++ src/main.cpp
@@ -3,21 +3,11 @@
 #endif
 #include "common_types.cpp"
 #include "system_state.cpp"
-#include "parsers.cpp"
-#include "defines.cpp"
 #include "float_from_chars.cpp"
-#include "gui_graphics_parsers.cpp"
 #include "text.cpp"
 #include "fonts.cpp"
 #include "texture.cpp"
-#include "nations_parsing.cpp"
-#include "cultures_parsing.cpp"
-#include "econ_parsing.cpp"
-#include "military_parsing.cpp"
 #include "date_interface.cpp"
-#include "provinces_parsing.cpp"
-#include "trigger_parsing.cpp"
-#include "effect_parsing.cpp"
 #include "serialization.cpp"
 #include "nations.cpp"
 #include "culture.cpp"
@@ -30,7 +20,6 @@
 #include "demographics.cpp"
 #include "bmfont.cpp"
 #include "rebels.cpp"
-#include "parsers_declarations.cpp"
 #include "politics.cpp"
 #include "events.cpp"
 #include "gui_graphics.cpp"
