+++ src/parsing/effect_parsing.cpp
@@ -1,5 +1,6 @@
 #include "effect_parsing.hpp"
 #include "parsers_declarations.hpp"
+#include "system_state.hpp"
 
 namespace parsers {
 
@@ -421,7 +422,7 @@ void ef_scope_all_core(token_generator& gen, error_handler& err, effect_building
 	context.limit_position = context.compiled_effect.size();
 	context.compiled_effect.push_back(trigger::payload(dcon::trigger_key()).value);
 
-	
+
 	parse_effect_body(gen, err, context);
 
 	context.compiled_effect[payload_size_offset] = uint16_t(context.compiled_effect.size() - payload_size_offset);
@@ -1201,30 +1202,7 @@ int32_t simplify_effect(uint16_t* source) {
 	}
 }
 
-template<typename T>
-void recurse_over_effects(uint16_t* source, T const& f) {
-	f(source);
-
-	if((source[0] & effect::code_mask) >= effect::first_scope_code) {
-		if((source[0] & effect::code_mask) == effect::random_list_scope) {
-			auto const source_size = 1 + effect::get_generic_effect_payload_size(source);
-
-			auto sub_units_start = source + 4; // [code] + [payload size] + [chances total] + [first sub effect chance]
-			while(sub_units_start < source + source_size) {
-				recurse_over_effects(sub_units_start, f);
-				sub_units_start += 2 + effect::get_generic_effect_payload_size(sub_units_start); // each member preceded by uint16_t
-			}
-		} else {
-			auto const source_size = 1 + effect::get_generic_effect_payload_size(source);
-
-			auto sub_units_start = source + 2 + effect::effect_scope_data_payload(source[0]);
-			while(sub_units_start < source + source_size) {
-				recurse_over_effects(sub_units_start, f);
-				sub_units_start += 1 + effect::get_generic_effect_payload_size(sub_units_start);
-			}
-		}
-	}
-}
+template void recurse_over_effects(uint16_t*, std::function<void(uint16_t*)> const&);
 
 dcon::effect_key make_effect(token_generator& gen, error_handler& err, effect_building_context& context) {
 	ef_scope_hidden_tooltip(gen, err, context);
