+++ src/parsing/effect_parsing.hpp
@@ -2661,7 +2661,7 @@ struct effect_body {
 			}
 			context.compiled_effect.push_back(trigger::payload(int16_t(value.value)).value);
 		} else if(value.who.length() == 3) {
-			
+
 			if(auto it = context.outer_context.map_of_ident_names.find(nations::tag_to_int(value.who[0], value.who[1], value.who[2]));
 					it != context.outer_context.map_of_ident_names.end()) {
 
@@ -3696,6 +3696,31 @@ void ef_scope_random_list(token_generator& gen, error_handler& err, effect_build
 void ef_scope_variable(std::string_view label, token_generator& gen, error_handler& err, effect_building_context& context);
 void ef_scope_any_substate(token_generator& gen, error_handler& err, effect_building_context& context);
 int32_t add_to_random_list(std::string_view label, token_generator& gen, error_handler& err, effect_building_context& context);
+
+template<typename T>
+void recurse_over_effects(uint16_t* source, T const& f) {
+	f(source);
+
+	if((source[0] & effect::code_mask) >= effect::first_scope_code) {
+		if((source[0] & effect::code_mask) == effect::random_list_scope) {
+			auto const source_size = 1 + effect::get_generic_effect_payload_size(source);
+
+			auto sub_units_start = source + 4; // [code] + [payload size] + [chances total] + [first sub effect chance]
+			while(sub_units_start < source + source_size) {
+				recurse_over_effects(sub_units_start, f);
+				sub_units_start += 2 + effect::get_generic_effect_payload_size(sub_units_start); // each member preceded by uint16_t
+			}
+		} else {
+			auto const source_size = 1 + effect::get_generic_effect_payload_size(source);
+
+			auto sub_units_start = source + 2 + effect::effect_scope_data_payload(source[0]);
+			while(sub_units_start < source + source_size) {
+				recurse_over_effects(sub_units_start, f);
+				sub_units_start += 1 + effect::get_generic_effect_payload_size(sub_units_start);
+			}
+		}
+	}
+}
 
 dcon::effect_key make_effect(token_generator& gen, error_handler& err, effect_building_context& context);
 int32_t simplify_effect(uint16_t* source);
