+++ src/parsing/military_parsing.cpp
@@ -1,5 +1,6 @@
 #include "parsers_declarations.hpp"
 #include "military.hpp"
+#include "system_state.hpp"
 
 namespace parsers {
